eta overflow fix
Sometime I get eta's like "eta 4294967286m:4294967266s" , just an overflow
issue. The following patch fixes the issue.

+++ eta.c
@@ -66,7 +66,7 @@ static void check_str_update(struct thread_data *td)
 /*
  * Convert seconds to a printable string.
  */
-static void eta_to_str(char *str, int eta_sec)
+static void eta_to_str(char *str, unsigned long eta_sec)
 {
 	unsigned int d, h, m, s;
 	int disp_hour = 0;
